+++ src/main/java/Party.java
@@ -4,13 +4,13 @@ import java.util.List;
 import java.util.Set;
 
 public class Party{
-  private Integer mGuests;
-  private Integer mFoodChoice; //0 = no food (g*0), 1 = light snacks(g*2), 2 = hd (g*3), 3 = pizza (g*4), 4 = meal (g*15)
-  private Integer mDrinkChoice; //0 = no drinks (g*0), 1 = tea & coffee (g*1), 2 = beer (g*5), 3 = bar (g*10)
-  private Integer mEntertainmentChoice; //0 = entertainment, 1 = DJ (+100), 2 = clowns (+30), 3 = Magician (+60), 4 = speaker (+200)
+  private double mGuests;
+  private double mFoodChoice; //0 = no food (g*0), 1 = light snacks(g*2), 2 = hd (g*3), 3 = pizza (g*4), 4 = meal (g*15)
+  private double mDrinkChoice; //0 = no drinks (g*0), 1 = tea & coffee (g*1), 2 = beer (g*5), 3 = bar (g*10)
+  private double mEntertainmentChoice; //0 = entertainment, 1 = DJ (+100), 2 = clowns (+30), 3 = Magician (+60), 4 = speaker (+200)
   private String mCoupon; // SUPERDEAL20 20%off, DJDISCOUNT30 $30off with DJ, MAGICALMAYHEM free Magician
 
-  public Party(Integer guest, Integer food, Integer drink, Integer entertainment, String coupons) {
+  public Party(double guest, double food, double drink, double entertainment, String coupons) {
     mGuests = guest;
     mFoodChoice = food;
     mDrinkChoice = drink;
@@ -18,7 +18,7 @@ public class Party{
     mCoupon = coupons;
   }
 
-  public Integer getGuests(){
+  public double getGuests(){
     return mGuests;
   }
   public String getFoodChoice(){
@@ -76,8 +76,8 @@ public class Party{
     return couponString;
   }
 
-  public Integer partyPrice (){
-    Integer price = 0;
+  public double partyPrice (){
+     double price = 0;
     price = mGuests*3;
     if (mFoodChoice == 0){
       price += mGuests*0;
